Send event as name
Backwards compatible with using event: true

This is actually a really large change, and requires updating quite a
lot. Right now, events are shown in the "pages" overview, but storing
"event" as string means we'll have to build a new overview. Quite a lot
of changes. Don't think this was a good idea.

Need a new plan for this.